requirements.txt
selenium
pillow
Selenium-Screenshot